requirements.txt
Flask
gunicorn
fastembed==0.5.1
tiktoken
requests
python-dotenv
qdrant-client